+++ src/ui/layout.rs
@@ -60,7 +60,7 @@ impl LayoutAreas {
             .vertical_margin(1)
             .constraints(
                 [
-                    Constraint::Length(5),       // line score
+                    Constraint::Length(4),       // line score
                     Constraint::Percentage(100), // box score
                 ]
                 .as_ref(),
